Built-in support for legacy mode (see desc)
If the user is using a theme with framework version prior to 2.5,
legacy mode will kick in. They’re forced to have the old three
sections: Featured, Main, Featured Below. There’s no new elements or
display settings. When working on the Edit Page screen, they can sync
the page to a template, but can’t edit the layout there.