Improvement: Allow layout to be empty (#14)
Why? Before, there always had to be at least one section for a layout
to work correctly. So the layout builder forces there to always have
that one section and new layouts are created with that section. That
one section becomes locked and it can’t be deleted. This means if you
started with a layout having only one section and you merge a template,
now you have an empty section you can’t delete.  This is why now
layouts can be empty and all sections can potentially be deleted.